// src/app/models/owner.ts
export class Owner {
    id?: number; 
    firstname?: string;
    lastname?: string;
    login?: string;
    password?: string; 
    phone?: string;
    etat?: string;
    email?: string;
    country?: string;
    city?: string;
    street?: string;
    codePostal?: string;
    image?: string; 
    loginType?: string;
    codeVerif?: string;
}